spec: deflake focus handling test
